# Conflicts solved:
#	macosx/EternityEngine.xcodeproj/project.pbxproj
#	midiproc/vc2008/midiproc.sln
#	source/d_iwad.cpp
#	source/e_args.cpp
#	source/e_sound.cpp
#	source/e_states.cpp
#	source/ev_sectors.cpp
#	source/p_floor.cpp
#	source/p_genlin.cpp
#	source/p_lights.cpp
#	source/p_partcl.cpp
#	source/p_plats.cpp
#	source/p_portal.cpp
#	source/p_scroll.cpp
#	source/p_sight.cpp
#	source/p_spec.cpp
#	source/p_switch.cpp
#	source/textscreen/txt_checkbox.c
#	source/textscreen/txt_dropdown.c
#	source/textscreen/txt_separator.c
#	vc2017/Eternity.vcxproj.filters

+++ source/w_wad.cpp
@@ -936,7 +936,8 @@ bool WadDirectory::addDirectoryAsArchive(openwad_t &openData,
       std::sort(paths.begin(), paths.end());
 
       // we now have the files neatly ordered
-      lumpinfo_t *lump_p = reAllocLumpInfo(paths.getLength(), startlump);
+      lumpinfo_t *lump_p = reAllocLumpInfo(static_cast<int>(paths.getLength()),
+                                           startlump);
 
       for(int i = startlump; i < numlumps; i++, lump_p++)
       {
